4.0.0-RC2 compatibility mybatis-plus pagination

+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/sharding/segment/select/pagination/Pagination.java
@@ -43,22 +43,22 @@ public final class Pagination {
     
     private final PaginationValueSegment rowCountSegment;
     
-    private final int actualOffset;
+    private final long actualOffset;
     
-    private final Integer actualRowCount;
+    private final Long actualRowCount;
     
     public Pagination(final PaginationValueSegment offsetSegment, final PaginationValueSegment rowCountSegment, final List<Object> parameters) {
         hasPagination = null != offsetSegment || null != rowCountSegment;
         this.offsetSegment = offsetSegment;
         this.rowCountSegment = rowCountSegment;
-        actualOffset = null == offsetSegment ? 0 : getValue(offsetSegment, parameters);
+        actualOffset = null == offsetSegment ? 0L : getValue(offsetSegment, parameters);
         actualRowCount = null == rowCountSegment ? null : getValue(rowCountSegment, parameters); 
     }
     
-    private int getValue(final PaginationValueSegment paginationValueSegment, final List<Object> parameters) {
+    private long getValue(final PaginationValueSegment paginationValueSegment, final List<Object> parameters) {
         return paginationValueSegment instanceof ParameterMarkerPaginationValueSegment
-                ? (int) parameters.get(((ParameterMarkerPaginationValueSegment) paginationValueSegment).getParameterIndex())
-                : ((NumberLiteralPaginationValueSegment) paginationValueSegment).getValue();
+                ? (long) parameters.get(((ParameterMarkerPaginationValueSegment) paginationValueSegment).getParameterIndex())
+                : (long) ((NumberLiteralPaginationValueSegment) paginationValueSegment).getValue();
     }
     
     /**
@@ -84,11 +84,11 @@ public final class Pagination {
      * 
      * @return actual offset
      */
-    public int getActualOffset() {
+    public long getActualOffset() {
         if (null == offsetSegment) {
-            return 0;
+            return 0L;
         }
-        return offsetSegment.isBoundOpened() ? actualOffset - 1 : actualOffset;
+        return offsetSegment.isBoundOpened() ? actualOffset - 1L : actualOffset;
     }
     
     /**
@@ -96,11 +96,11 @@ public final class Pagination {
      *
      * @return actual row count
      */
-    public Optional<Integer> getActualRowCount() {
+    public Optional<Long> getActualRowCount() {
         if (null == rowCountSegment) {
             return Optional.absent();
         }
-        return Optional.of(rowCountSegment.isBoundOpened() ? actualRowCount + 1 : actualRowCount);
+        return Optional.of(rowCountSegment.isBoundOpened() ? actualRowCount + 1L : actualRowCount);
     }
     
     /**
@@ -127,8 +127,8 @@ public final class Pagination {
      *
      * @return revised offset
      */
-    public int getRevisedOffset() {
-        return 0;
+    public long getRevisedOffset() {
+        return 0L;
     }
     
     /**
@@ -137,9 +137,9 @@ public final class Pagination {
      * @param optimizedStatement optimized statement
      * @return revised row count
      */
-    public int getRevisedRowCount(final ShardingSelectOptimizedStatement optimizedStatement) {
+    public long getRevisedRowCount(final ShardingSelectOptimizedStatement optimizedStatement) {
         if (isMaxRowCount(optimizedStatement)) {
-            return Integer.MAX_VALUE;
+            return Long.MAX_VALUE;
         }
         return rowCountSegment instanceof LimitValueSegment ? actualOffset + actualRowCount : actualRowCount;
     }
